services/api-gateway/requirements.txt
# --- Web Framework ---
fastapi==0.121.1
uvicorn[standard]==0.38.0
httpx==0.28.1

# --- OpenTelemetry (OTel) ---
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0

# --- OTel Instrumentation ---
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
opentelemetry-instrumentation-logging==0.57b0

# --- OTel Exporter ---
opentelemetry-exporter-otlp-proto-grpc==1.36.0
